+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Lallemand 2023 benzylpenicillin ([doi:10.3389/fmicb.2023.1282949](https://doi.org/10.3389/fmicb.2023.1282949)) - healthy horses.
+
 - Add Wu 2023 SPI-62 ([doi:10.1007/s40262-023-01278-8](https://doi.org/10.1007/s40262-023-01278-8)) - healthy adults.
 
 - Disambiguated the overloaded `OC` name, which denoted five unrelated